Refactor duplicated codes
